A simple encryption and decryption software
Client-Server software where a client can send a post to send to a server. It will be encrypted if they want it to. The server checks a signature received and will encrypt or post the message. On connection the software will decrypt messages if the user who joins is the recipient of a message.